refactor(wasm): split c file
BREAKING CHANGE: reinstall project dependencies
