+++ apps/webaudio-spectrum-analyser/src/app/wasm/fft.h
@@ -1,0 +1,6 @@
+#ifndef FFT_H_INCLUDED
+#define FFT_H_INCLUDED
+
+#include "common.h"
+
+#endif
