+++ apps/webaudio-spectrum-analyser/src/app/wasm/index.c
@@ -1,0 +1,4 @@
+#include "common.c"
+#include "autocorr.c"
+#include "prominence.c"
+#include "fft.c"
