+++ jest.config.js
@@ -10,7 +10,7 @@ function getWasmOptions() {
       rules: [],
     },
   });
-  return cfg.module.rules[0].use.options;
+  return cfg.module.rules[0].use[0].options;
 }
 
 module.exports = {
