+++ package.json
@@ -132,6 +132,7 @@
     "babel-loader": "8.1.0",
     "codelyzer": "6.0.0",
     "commitizen": "4.2.1",
+    "cpp-dependency-loader": "^0.1.0",
     "cpp-wasm-loader": "0.9.0",
     "cypress": "5.2.0",
     "cypress-multi-reporters": "1.4.0",
